implementar Interfaz SueldoNetoSegunCategoria

+++ SueldoAvanzado.java
@@ -2,7 +2,7 @@ package SuledoPepe;
 
 public class SueldoAvanzado extends CalcularSueldo{
 
-    public float calcular(Persona empleado){
+    public double calcular(Persona empleado){
         return 0.0F;
     }
 }
